+++ src/Server/ServerCore/Systems/Scripting/ScriptingController.cs
@@ -132,15 +132,19 @@ public class ScriptingController(ScriptManager scriptManager, ILogger<ScriptingC
 
     /// <summary>
     ///     Requests that the interaction callback of the target entity be invoked with the given tool entity.
-    ///     If the target has an interaction callback, it is invoked with the tool entity ID as the first Lua argument.
+    ///     If the target has an interaction callback, it is invoked with the using entity ID, tool entity ID,
+    ///     and target entity ID as arguments, in that order.
     /// </summary>
     /// <param name="eventSender">Event sender.</param>
+    /// <param name="usingEntityId">Entity ID of the entity using the tool.</param>
     /// <param name="toolEntityId">Tool entity ID.</param>
     /// <param name="targetEntityId">Target entity ID.</param>
-    public void InvokeInteractCallback(IEventSender eventSender, ulong toolEntityId, ulong targetEntityId)
+    public void InvokeInteractCallback(IEventSender eventSender, ulong usingEntityId, ulong toolEntityId,
+        ulong targetEntityId)
     {
         var details = new ScriptingInteractEventDetails
         {
+            UsingEntityId = usingEntityId,
             ToolEntityId = toolEntityId,
             TargetEntityId = targetEntityId
         };