+++ src/Server/ServerCore/Systems/Scripting/ScriptingSystem.cs
@@ -18,13 +18,11 @@ using System;
 using System.Collections.Generic;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Sovereign.EngineCore.Entities;
 using Sovereign.EngineCore.Events;
 using Sovereign.EngineCore.Events.Details;
 using Sovereign.EngineCore.Events.Details.Validators;
 using Sovereign.EngineCore.Lua;
 using Sovereign.EngineCore.Systems;
-using Sovereign.EngineCore.Systems.Data;
 using Sovereign.ServerCore.Configuration;
 using EventId = Sovereign.EngineCore.Events.EventId;
 
@@ -36,13 +34,11 @@ namespace Sovereign.ServerCore.Systems.Scripting;
 internal class ScriptingSystem : ISystem
 {
     private readonly ScriptingCallbackManager callbackManager;
-    private readonly IDataServices dataServices;
     private readonly EntityCallbacks entityCallbacks;
     private readonly EntityScriptCallbacks entityScriptCallbacks;
     private readonly ILogger<ScriptingSystem> logger;
     private readonly ScriptManager manager;
     private readonly ScriptLoader scriptLoader;
-    private readonly ScriptingController scriptingController;
     private readonly TestHarnessResultsCollector testHarnessCollector;
     private readonly TestHarnessOptions testHarnessOptions;
     private readonly ITimedCallbackRunner timedCallbackRunner;
@@ -51,8 +47,7 @@ internal class ScriptingSystem : ISystem
         ScriptingCallbackManager callbackManager, ScriptLoader scriptLoader, ILogger<ScriptingSystem> logger,
         EntityScriptCallbacks entityScriptCallbacks, ITimedCallbackRunner timedCallbackRunner,
         EntityCallbacks entityCallbacks, TestHarnessResultsCollector testHarnessCollector,
-        IOptions<TestHarnessOptions> testHarnessOptions, IDataServices dataServices,
-        ScriptingController scriptingController)
+        IOptions<TestHarnessOptions> testHarnessOptions)
     {
         this.manager = manager;
         this.callbackManager = callbackManager;
@@ -63,8 +58,6 @@ internal class ScriptingSystem : ISystem
         this.entityCallbacks = entityCallbacks;
         this.testHarnessCollector = testHarnessCollector;
         this.testHarnessOptions = testHarnessOptions.Value;
-        this.dataServices = dataServices;
-        this.scriptingController = scriptingController;
         EventCommunicator = eventCommunicator;
 
         EventIdsOfInterest = new HashSet<EventId>(ScriptableEventSet.Events);
@@ -146,7 +139,8 @@ internal class ScriptingSystem : ISystem
                         break;
                     }
 
-                    OnInteractCallback(details);
+                    entityScriptCallbacks.InvokeInteractCallback(details.UsingEntityId, details.ToolEntityId,
+                        details.TargetEntityId);
                     break;
                 }
 
@@ -334,20 +328,5 @@ internal class ScriptingSystem : ISystem
     private void OnTimedCallback(IntPtr luaState, int callbackRef, int argRef)
     {
         timedCallbackRunner.RunTimedCallback(luaState, callbackRef, argRef);
-    }
-
-    /// <summary>
-    ///     Called when an InteractCallback event is received.
-    /// </summary>
-    /// <param name="details">Event details.</param>
-    private void OnInteractCallback(ScriptingInteractEventDetails details)
-    {
-        if (!dataServices.TryGetEntityKeyValue(details.TargetEntityId, EntityConstants.InteractScriptKey,
-                out var scriptName) ||
-            !dataServices.TryGetEntityKeyValue(details.TargetEntityId, EntityConstants.InteractFunctionKey,
-                out var funcName)) return;
-
-        scriptingController.CallFunctionAsync(scriptName, funcName, details.ToolEntityId,
-            details.TargetEntityId);
     }
 }